requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.8
open-clip-torch>=2.24.0
numpy>=1.26.0
pillow>=10.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
